refactor : 해당 클래스들 application/usecase 폴더 밑으로 이동
